+++ test/ERC20TokenShares.t.sol
@@ -18,7 +18,7 @@ contract ERC20TokenSharesTest is Test {
     function testERC20Decimals() public {
         address tokenizedShares;
         address[] memory recipients;
-        uint256[] memory shares;
+        uint16[] memory shares;
 
         // Success
         recipients = new address[](3);
@@ -26,7 +26,7 @@ contract ERC20TokenSharesTest is Test {
         recipients[1] = makeAddr("recipient_1");
         recipients[2] = makeAddr("recipient_2");
 
-        shares = new uint256[](3);
+        shares = new uint16[](3);
         shares[0] = 7_000;
         shares[1] = 2_000;
         shares[2] = 1_000;
@@ -44,7 +44,7 @@ contract ERC20TokenSharesTest is Test {
     function testERC20AddTokenizedShares() public {
         address tokenizedShares;
         address[] memory recipients;
-        uint256[] memory shares;
+        uint16[] memory shares;
 
         // Revert ITokenizedShares__NoRecipients
         recipients = new address[](0);
@@ -57,12 +57,12 @@ contract ERC20TokenSharesTest is Test {
         recipients[1] = makeAddr("recipient_1");
         recipients[2] = makeAddr("recipient_2");
 
-        shares = new uint256[](1);
+        shares = new uint16[](1);
         vm.expectRevert(ITokenizedShares.ITokenizedShares__ArrayLengthsMismatch.selector);
         factory.addTokenizedShares(recipients, shares);
 
         // Revert ITokenizedShares__InvalidSharesAmount
-        shares = new uint256[](3);
+        shares = new uint16[](3);
         shares[0] = 6_999;
         shares[1] = 2_000;
         shares[2] = 1_000;
@@ -94,9 +94,9 @@ contract ERC20TokenSharesTest is Test {
 
     function testERC20AddTokenizedSharesWithKeeperShares() public {
         address tokenizedShares;
-        uint256 keeperShares;
+        uint16 keeperShares;
         address[] memory recipients;
-        uint256[] memory shares;
+        uint16[] memory shares;
 
         recipients = new address[](3);
         recipients[0] = makeAddr("recipient_0");
@@ -105,7 +105,7 @@ contract ERC20TokenSharesTest is Test {
 
         // Revert ITokenizedShares__InvalidSharesAmount
         keeperShares = 110;
-        shares = new uint256[](3);
+        shares = new uint16[](3);
         shares[0] = 6_889;
         shares[1] = 2_000;
         shares[2] = 1_000;
@@ -134,9 +134,9 @@ contract ERC20TokenSharesTest is Test {
     function testERC20AddTokenizedSharesWithKeeperSharesithCustomImplementation() public {
         address customImplementation;
         address tokenizedShares;
-        uint256 keeperShares;
+        uint16 keeperShares;
         address[] memory recipients;
-        uint256[] memory shares;
+        uint16[] memory shares;
 
         recipients = new address[](3);
         recipients[0] = makeAddr("recipient_0");
@@ -145,7 +145,7 @@ contract ERC20TokenSharesTest is Test {
 
         // Revert ITokenizedShares__InvalidSharesAmount
         keeperShares = 110;
-        shares = new uint256[](3);
+        shares = new uint16[](3);
         shares[0] = 6_889;
         shares[1] = 2_000;
         shares[2] = 1_000;
@@ -175,13 +175,13 @@ contract ERC20TokenSharesTest is Test {
     function testERC20MultiAddTokenizedShares() public {
         uint256 number = 10;
 
-        for (uint256 i = 1; i <= number; ++i) {
+        for (uint16 i = 1; i <= number; ++i) {
             address[] memory recipients = new address[](3);
             recipients[0] = makeAddr(string(abi.encode(i)));
             recipients[1] = makeAddr(string(abi.encode(i + i)));
             recipients[2] = makeAddr(string(abi.encode(i + i + i)));
 
-            uint256[] memory shares = new uint256[](3);
+            uint16[] memory shares = new uint16[](3);
             shares[0] = 7_000;
             shares[1] = 2_000 + i;
             shares[2] = 1_000 - i;
@@ -205,14 +205,14 @@ contract ERC20TokenSharesTest is Test {
     function testERC20MultiAddTokenizedSharesWithKeeperShares() public {
         uint256 number = 10;
 
-        for (uint256 i = 1; i <= number; ++i) {
+        for (uint16 i = 1; i <= number; ++i) {
             address[] memory recipients = new address[](3);
             recipients[0] = makeAddr(string(abi.encode(i)));
             recipients[1] = makeAddr(string(abi.encode(i + i)));
             recipients[2] = makeAddr(string(abi.encode(i + i + i)));
 
-            uint256 keeperShares = 111;
-            uint256[] memory shares = new uint256[](3);
+            uint16 keeperShares = 111;
+            uint16[] memory shares = new uint16[](3);
             shares[0] = 6_889;
             shares[1] = 2_000 + i;
             shares[2] = 1_000 - i;
@@ -241,7 +241,7 @@ contract ERC20TokenSharesTest is Test {
         recipients[1] = makeAddr("recipient_1");
         recipients[2] = makeAddr("recipient_2");
 
-        uint256[] memory shares = new uint256[](3);
+        uint16[] memory shares = new uint16[](3);
         shares[0] = 7_000;
         shares[1] = 2_000;
         shares[2] = 1_000;
@@ -265,8 +265,8 @@ contract ERC20TokenSharesTest is Test {
         recipients[1] = makeAddr("recipient_1");
         recipients[2] = makeAddr("recipient_2");
 
-        uint256 keeperShares = 111;
-        uint256[] memory shares = new uint256[](3);
+        uint16 keeperShares = 111;
+        uint16[] memory shares = new uint16[](3);
         shares[0] = 6_889;
         shares[1] = 2_000;
         shares[2] = 1_000;
@@ -291,7 +291,7 @@ contract ERC20TokenSharesTest is Test {
         recipients[1] = makeAddr("recipient_1");
         recipients[2] = makeAddr("recipient_2");
 
-        uint256[] memory shares = new uint256[](3);
+        uint16[] memory shares = new uint16[](3);
         shares[0] = 7_000;
         shares[1] = 2_000;
         shares[2] = 1_000;
@@ -326,8 +326,8 @@ contract ERC20TokenSharesTest is Test {
         recipients[1] = makeAddr("recipient_1");
         recipients[2] = makeAddr("recipient_2");
 
-        uint256 keeperShares = 111;
-        uint256[] memory shares = new uint256[](3);
+        uint16 keeperShares = 111;
+        uint16[] memory shares = new uint16[](3);
         shares[0] = 6_889;
         shares[1] = 2_000;
         shares[2] = 1_000;
@@ -369,7 +369,7 @@ contract ERC20TokenSharesTest is Test {
         recipients[1] = makeAddr("recipient_1");
         recipients[2] = makeAddr("recipient_2");
 
-        uint256[] memory shares = new uint256[](3);
+        uint16[] memory shares = new uint16[](3);
         shares[0] = 7_000;
         shares[1] = 2_000;
         shares[2] = 1_000;
@@ -407,7 +407,7 @@ contract ERC20TokenSharesTest is Test {
     function testERC20ReleasableWithTransfers() public {
         uint256 amount = 0.000000001 ether;
         address[] memory recipients;
-        uint256[] memory shares;
+        uint16[] memory shares;
         bool success;
 
         address recipient_0 = makeAddr("recipient_0");
@@ -421,7 +421,7 @@ contract ERC20TokenSharesTest is Test {
         recipients[1] = recipient_1;
         recipients[2] = recipient_2;
 
-        shares = new uint256[](3);
+        shares = new uint16[](3);
         shares[0] = 7_000;
         shares[1] = 2_000;
         shares[2] = 1_000;
@@ -475,7 +475,7 @@ contract ERC20TokenSharesTest is Test {
     function testERC20ReleasableWithTransfersAfterRelease() public {
         uint256 amount = 0.001 ether;
         address[] memory recipients;
-        uint256[] memory shares;
+        uint16[] memory shares;
         bool success;
 
         address recipient_0 = makeAddr("recipient_0");
@@ -489,7 +489,7 @@ contract ERC20TokenSharesTest is Test {
         recipients[1] = recipient_1;
         recipients[2] = recipient_2;
 
-        shares = new uint256[](3);
+        shares = new uint16[](3);
         shares[0] = 7_000;
         shares[1] = 2_000;
         shares[2] = 1_000;
@@ -555,19 +555,19 @@ contract ERC20TokenSharesTest is Test {
     //                FUZZ                  //
     //--------------------------------------//
 
-    function testERC20FuzzAddTokenizedShares(uint256 shares1, uint256 shares2) public {
+    function testERC20FuzzAddTokenizedShares(uint16 shares1, uint16 shares2) public {
         vm.assume(shares1 > 0 && shares1 < 10_000);
         vm.assume(shares2 > 0 && shares2 < 10_000);
         vm.assume(shares1 + shares2 < 10_000);
 
-        uint256 shares3 = 10_000 - shares1 - shares2;
+        uint16 shares3 = 10_000 - shares1 - shares2;
 
         address[] memory recipients = new address[](3);
         recipients[0] = makeAddr("recipient_0");
         recipients[1] = makeAddr("recipient_1");
         recipients[2] = makeAddr("recipient_2");
 
-        uint256[] memory shares = new uint256[](3);
+        uint16[] memory shares = new uint16[](3);
         shares[0] = shares1;
         shares[1] = shares2;
         shares[2] = shares3;
@@ -583,7 +583,7 @@ contract ERC20TokenSharesTest is Test {
         assertEq(ERC20(tokenizedShares).balanceOf(recipients[2]), shares[2] * 10 ** decimals);
     }
 
-    function testERC20FuzzAddTokenizedSharesWithKeeperShares(uint256 keeperShares, uint256 shares1, uint256 shares2)
+    function testERC20FuzzAddTokenizedSharesWithKeeperShares(uint16 keeperShares, uint16 shares1, uint16 shares2)
         public
     {
         vm.assume(keeperShares > 0 && keeperShares <= 1_000);
@@ -591,14 +591,14 @@ contract ERC20TokenSharesTest is Test {
         vm.assume(shares2 > 0 && shares2 < 10_000);
         vm.assume(shares1 + shares2 + keeperShares < 10_000);
 
-        uint256 shares3 = 10_000 - shares1 - shares2 - keeperShares;
+        uint16 shares3 = 10_000 - shares1 - shares2 - keeperShares;
 
         address[] memory recipients = new address[](3);
         recipients[0] = makeAddr("recipient_0");
         recipients[1] = makeAddr("recipient_1");
         recipients[2] = makeAddr("recipient_2");
 
-        uint256[] memory shares = new uint256[](3);
+        uint16[] memory shares = new uint16[](3);
         shares[0] = shares1;
         shares[1] = shares2;
         shares[2] = shares3;
@@ -616,23 +616,23 @@ contract ERC20TokenSharesTest is Test {
     }
 
     function testERC20FuzzAddTokenizedSharesWithKeeperSharesWithCustomImplementation(
-        uint256 keeperShares,
-        uint256 shares1,
-        uint256 shares2
+        uint16 keeperShares,
+        uint16 shares1,
+        uint16 shares2
     ) public {
         vm.assume(keeperShares > 0 && keeperShares <= 1_000);
         vm.assume(shares1 > 0 && shares1 < 10_000);
         vm.assume(shares2 > 0 && shares2 < 10_000);
         vm.assume(shares1 + shares2 + keeperShares < 10_000);
 
-        uint256 shares3 = 10_000 - shares1 - shares2 - keeperShares;
+        uint16 shares3 = 10_000 - shares1 - shares2 - keeperShares;
 
         address[] memory recipients = new address[](3);
         recipients[0] = makeAddr("recipient_0");
         recipients[1] = makeAddr("recipient_1");
         recipients[2] = makeAddr("recipient_2");
 
-        uint256[] memory shares = new uint256[](3);
+        uint16[] memory shares = new uint16[](3);
         shares[0] = shares1;
         shares[1] = shares2;
         shares[2] = shares3;
@@ -650,22 +650,22 @@ contract ERC20TokenSharesTest is Test {
         assertEq(ERC20(tokenizedShares).balanceOf(recipients[2]), shares[2] * 10 ** decimals);
     }
 
-    function testERC20FuzzMultiAddTokenizedShares(uint256 shares1, uint256 shares2) public {
+    function testERC20FuzzMultiAddTokenizedShares(uint16 shares1, uint16 shares2) public {
         vm.assume(shares1 > 0 && shares1 < 10_000);
         vm.assume(shares2 > 0 && shares2 < 10_000);
         vm.assume(shares1 + shares2 < 10_000);
 
-        uint256 shares3 = 10_000 - shares1 - shares2;
+        uint16 shares3 = 10_000 - shares1 - shares2;
         uint256 number = 10;
         vm.assume(shares3 > number);
 
-        for (uint256 i = 1; i <= number; ++i) {
+        for (uint16 i = 1; i <= number; ++i) {
             address[] memory recipients = new address[](3);
             recipients[0] = makeAddr(string(abi.encode(i)));
             recipients[1] = makeAddr(string(abi.encode(i + i)));
             recipients[2] = makeAddr(string(abi.encode(i + i + i)));
 
-            uint256[] memory shares = new uint256[](3);
+            uint16[] memory shares = new uint16[](3);
             shares[0] = shares1;
             shares[1] = shares2 + i;
             shares[2] = shares3 - i;
@@ -681,27 +681,25 @@ contract ERC20TokenSharesTest is Test {
         }
     }
 
-    function testERC20FuzzMultiAddTokenizedSharesWithKeeperShares(
-        uint256 keeperShares,
-        uint256 shares1,
-        uint256 shares2
-    ) public {
+    function testERC20FuzzMultiAddTokenizedSharesWithKeeperShares(uint16 keeperShares, uint16 shares1, uint16 shares2)
+        public
+    {
         vm.assume(keeperShares > 0 && keeperShares <= 1_000);
         vm.assume(shares1 > 0 && shares1 < 10_000);
         vm.assume(shares2 > 0 && shares2 < 10_000);
         vm.assume(shares1 + shares2 + keeperShares < 10_000);
 
-        uint256 shares3 = 10_000 - shares1 - shares2 - keeperShares;
+        uint16 shares3 = 10_000 - shares1 - shares2 - keeperShares;
         uint256 number = 10;
         vm.assume(shares3 > number);
 
-        for (uint256 i = 1; i <= number; ++i) {
+        for (uint16 i = 1; i <= number; ++i) {
             address[] memory recipients = new address[](3);
             recipients[0] = makeAddr(string(abi.encode(i)));
             recipients[1] = makeAddr(string(abi.encode(i + i)));
             recipients[2] = makeAddr(string(abi.encode(i + i + i)));
 
-            uint256[] memory shares = new uint256[](3);
+            uint16[] memory shares = new uint16[](3);
             shares[0] = shares1;
             shares[1] = shares2 + i;
             shares[2] = shares3 - i;
@@ -727,7 +725,7 @@ contract ERC20TokenSharesTest is Test {
         recipients[1] = makeAddr("recipient_1");
         recipients[2] = makeAddr("recipient_2");
 
-        uint256[] memory shares = new uint256[](3);
+        uint16[] memory shares = new uint16[](3);
         shares[0] = 7_000;
         shares[1] = 2_000;
         shares[2] = 1_000;
@@ -751,8 +749,8 @@ contract ERC20TokenSharesTest is Test {
         recipients[1] = makeAddr("recipient_1");
         recipients[2] = makeAddr("recipient_2");
 
-        uint256 keeperShares = 111;
-        uint256[] memory shares = new uint256[](3);
+        uint16 keeperShares = 111;
+        uint16[] memory shares = new uint16[](3);
         shares[0] = 6_889;
         shares[1] = 2_000;
         shares[2] = 1_000;
@@ -777,7 +775,7 @@ contract ERC20TokenSharesTest is Test {
         recipients[1] = makeAddr("recipient_1");
         recipients[2] = makeAddr("recipient_2");
 
-        uint256[] memory shares = new uint256[](3);
+        uint16[] memory shares = new uint16[](3);
         shares[0] = 7_000;
         shares[1] = 2_000;
         shares[2] = 1_000;
@@ -804,8 +802,8 @@ contract ERC20TokenSharesTest is Test {
         recipients[1] = makeAddr("recipient_1");
         recipients[2] = makeAddr("recipient_2");
 
-        uint256 keeperShares = 111;
-        uint256[] memory shares = new uint256[](3);
+        uint16 keeperShares = 111;
+        uint16[] memory shares = new uint16[](3);
         shares[0] = 6_889;
         shares[1] = 2_000;
         shares[2] = 1_000;
